+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - [8470](https://github.com/vegaprotocol/vega/issues/8470) - Stop orders snapshots
 - [8548](https://github.com/vegaprotocol/vega/issues/8548) - Use default for tendermint `RPC` address and better validation for `semver`
 - [8472](https://github.com/vegaprotocol/vega/issues/8472) - Add support for stop orders with batch market instructions
+- [8567](https://github.com/vegaprotocol/vega/issues/8567) - Add virtual stake and market growth to market data.
 
 ### 🐛 Fixes